Extract EditContent and add tests
Move EditContent into a file from todo_file::mod and add tests.
