+++ src/edit/mod.rs
@@ -4,7 +4,8 @@ use crate::input::Input;
 use crate::process::process_module::ProcessModule;
 use crate::process::process_result::ProcessResult;
 use crate::process::state::State;
-use crate::todo_file::{EditContext, TodoFile};
+use crate::todo_file::edit_content::EditContext;
+use crate::todo_file::TodoFile;
 use crate::view::line_segment::LineSegment;
 use crate::view::view_data::ViewData;
 use crate::view::view_line::ViewLine;
