+++ src/list/mod.rs
@@ -9,8 +9,9 @@ use crate::process::process_module::ProcessModule;
 use crate::process::process_result::ProcessResult;
 use crate::process::state::State;
 use crate::todo_file::action::Action;
+use crate::todo_file::edit_content::EditContext;
 use crate::todo_file::line::Line;
-use crate::todo_file::{EditContext, TodoFile};
+use crate::todo_file::TodoFile;
 use crate::view::view_data::ViewData;
 use crate::view::view_line::ViewLine;
 use crate::view::View;
